LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-return-value-expected.txt


FAIL Tests dialog.returnValue is settable and returns the last value set. assert_equals: expected "Return value set from close()" but got "7"
